isolates and moves manifest package to root mets package.
