+++ src/main/java/nl/kb/dare/endpoints/OaiRecordsEndpoint.java
@@ -19,8 +19,8 @@ import nl.kb.dare.model.statuscodes.ErrorStatus;
 import nl.kb.dare.model.statuscodes.OaiStatus;
 import nl.kb.dare.model.statuscodes.ProcessStatus;
 import nl.kb.dare.oai.GetRecord;
-import nl.kb.dare.manifest.ManifestXmlHandler;
-import nl.kb.dare.manifest.ObjectResource;
+import nl.kb.mets.manifest.ManifestXmlHandler;
+import nl.kb.mets.manifest.ObjectResource;
 import nl.kb.dare.xslt.XsltTransformer;
 import org.glassfish.jersey.server.ChunkedOutput;
 import org.skife.jdbi.v2.DBI;
