+++ src/main/java/nl/kb/mets/manifest/ObjectResource.java
@@ -1,21 +1,13 @@
-package nl.kb.dare.manifest;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
+package nl.kb.mets.manifest;
 
 // Represents a file node in Manifest, used to download and ship files with
 public class ObjectResource {
 
-    @JsonProperty
     private String id;
-    @JsonProperty
     private String xlinkHref;
-    @JsonProperty
     private String checksum;
-    @JsonProperty
     private String checksumType;
-    @JsonProperty
     private String localFilename;
-    @JsonProperty
     private long size;
 
     public String getId() {
