+++ src/test/java/nl/kb/dare/integration/IntegrationTest.java
@@ -10,8 +10,8 @@ import nl.kb.dare.model.oai.OaiRecord;
 import nl.kb.dare.model.repository.Repository;
 import nl.kb.dare.model.statuscodes.OaiStatus;
 import nl.kb.dare.model.statuscodes.ProcessStatus;
-import nl.kb.dare.manifest.ManifestXmlHandler;
-import nl.kb.dare.manifest.ObjectResource;
+import nl.kb.mets.manifest.ManifestXmlHandler;
+import nl.kb.mets.manifest.ObjectResource;
 import nl.kb.dare.oai.ScheduledOaiHarvester;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
