+++ src/test/java/nl/kb/dare/oai/GetRecordTest.java
@@ -2,7 +2,7 @@ package nl.kb.dare.oai;
 
 import com.google.common.collect.Lists;
 import nl.kb.filestorage.FileStorageHandle;
-import nl.kb.dare.manifest.ObjectResource;
+import nl.kb.mets.manifest.ObjectResource;
 import nl.kb.dare.model.oai.OaiRecord;
 import nl.kb.dare.model.reporting.ProgressReport;
 import nl.kb.dare.model.statuscodes.ProcessStatus;
